Add a file safe store to the collector
This safe will be used by the compliance framework to
distribute reference files in a secure way.

A node's shared secret will authent the download request,
and the node will be granted access to the safe files published
to its team responsible.

The files are stored in init/uploads/safe/ as hashed names.
Each file has a reference in the database, in the safe table:

+---------------+--------------+
| Field         | Type         |
+---------------+--------------+
| id            | int(11)      |
| uploader      | int(11)      |
| uploaded_from | varchar(512) |
| uploaded_date | datetime     |
| name          | varchar(512) |
| size          | int(11)      |
| uuid          | varchar(512) |
+---------------+--------------+

Two tables store the files publications and responsibles to
organizational groups:

- safe_team_publication
- safe_team_responsible

The rest api to manage the safe is:

/safe/upload
=> POST: upload a file, a symbolic name is welcome.
         the file is accessible and manageable only by its uploader

/safe
=> GET: list the files the user can access in the store

/safe/<file>
=> POST: change file metadata (ie, only the symbolic name as the other
properties are auditable and considered read-only)

/safe/<file>/download
=> GET: stream as multi-part doc

/safe/<file>/publications
/safe/<file>/responsibles
=> GET: list groups for role

/safe/<file>/publications/<group>
/safe/<file>/responsibles/<group>
=> POST: assign role to group
=> DELETE: unassign role to group

The cli is updated to support upload-from-file and download-to-file.