Disallow a node to set its team_responsible
Ruleset, modulesets and safe files are exposed to nodes with
selected team_responsibles. If the node can choose its
team_responsible it can gain access to content it should not
see.

This patch changes the xmlrpc and rest handlers to disallow
the nodes to change their team_responsible by removing the
key from the update data from the node.