Make services_ismemberof() an async + callback function
services_ismemberof() has not garanty the _groups cache is already
loaded when called. So it should not behave as a synchronous
function.

This patch makes it async, waiting for _groups to be initialized,
then executing the callback only if the user is member of the
required group passed as parameter.

Teach the sysreport to use that when deciding to display the
management tools.