Allow node properties edition directly from the ad-hoc tab
This first pass implements the "string" updater. The "date",
"datetime", "group", "app" and "integer" updater are yet to do.

Only the sys responsibles are presented with the updaters.

The updaters post data through the rest api.

This will lead to the suppression of the node_edit controller
fonction and view.