Search widget audit
1/ replace the "keys" and "key_separator" data key with "title"

keys and key_separator were only used to format the title.
set the title as a format string instead, with substitution
wildcard. Which let us be more creative in the titles and
permits not include all returned keys in the title.

2/ change the "tab" data key values

tab was a boolean. Now tab is a string representing either the
url to ajax-load, or the js function to call. In both cases, the
string can embed substitution wildcards. A special wildcard is
__rowid__, replaced by the id of the extrarow div.

3/ change the "links" data key values

links was expected to be always appended the same key value from
the search request resultset. This is not versatile enough. For
example, we want to pass the filterset id to the designer link.

4/ stop formatting html as strings

Use jquery objects for that, which helps maintenability.

5/ remove trailing whitespaces, indentation incoherence, ...

6/ remove the link to svcmon_log from filterset matches

7/ add node tabs to vm matches