Fix menu entry focus
When the nav menu is closed with an entry focused, the next menu
open leaves the selected entry unchanged. The <tab> then
selects the first menu entry, which causes 2 entries to be
selected. The arrow keys navigation in the menu then moves both
selections.

Fix by unselecting entries on open and close.