Add the /compliance/rulesets_variables POST handler
This handler is useful to create variables from form outputs, where
the ruleset id or name is not known before form submit. In this
case, having the ruleset id in the url path is not practical,
but a mangler can still pass it from the previous outputs result
into the data payload.