Move the rq tasks out of the web2py scheduler
os.kill() from inside the subprocess is not a good idea, but
we have no other way to exit the subprocess without hanging on
futex.

So run the rq tasks outside of the scheduler.

As a side gain, we now have meaningful process names, so the
resource usage per redis queue is more readable.