Make forms rest output and manglers server-side
So that we can PUT /forms/<id> and beneficit from these neat
features.

Mangling server-side need nodejs and its vm2 module, as the manglers
are in js and may contain untrusted/unreviewed code. As a consequence
new docker images need to be prepared if applicable. If nodejs and
vm2 are installed in a custom location, the new collector config
parameters "vm2" and "nodejs" allow specifying the paths.

Also spit the server side form submit function into per-output type
functions.