Support relaying rest requests
When the collector receives a request and the handler is flagged
for relaying it relays the request to the proxy collector in
charge of the node concerned by the request.

When a list-type payload is received, each list entry can be
relayed or not depending on which collector is front-facing the
node.

For a service with nodes attached to different collectors, the
request is relayed to each collector.

The handler can also be tagged for local treatment of the request
in addition to relaying, and for remote collector resync.

All handlers inherit the "local" treatment flag from their parent
class.

With this patch, only the tag attachments handlers are flagged
for relaying. Subsequent patches will activate relaying to all
object-to-node|service attachments.

Relaying supports arbitrary sized collector chains. The request
is handled first by the deeper collector and by the end-user
facing collector last.

Errors propagate from bottom to top.