Fix "agent actions > on services instances" displayed in the services view
No node_id available there, so this menu entry should not be
displayed.

This bug was due to the cache_id key not differentiating on
the selector condition, so that the cache entry for a selector
using the "node_id,svc_id" condition was fetched for selectors
using the "node_id+svc_id" condition.