Don't stack in PUT /form_output_results
When the results_id is not found.

Stop storing stdout and stderr in the script form output, as
they are already in the results log.

Pass the output_id and results to the from script as argv2 and
argv3, so the script can use PUT /form_output_results to
store results.