Enable/disable the body scroll while header zones are open
So the body scroll is locked even if the user continues scrolling
past the header zone bottom or up the header zone top.

This also prevents the double scroller effect.

The main user experience issue fixed is the accidental movement
of the body information while scrolling in a object tab opened
in the header zone.