Move the data management tools that don't use table data to the nav menu
Out of the table action menus, so that there access is more natural
and pervasive.

We also benefit from the nav menu filtering capabilities and
the hidden group menu entries system.

Another side benefit is that these tools now have a direct
access url and a js object usable in reports.

This patch also introduce the notion of internal forms. These
forms are not stored in database, so their updates will come
naturally with the collector codebase. Their id is negative, to
not conflict with the user-defined forms stored in db.

The internal forms are provided to the js clients through the
initial user dump rest GET call.