Cancel to focus autocomplete events in the form js object
Some custom focus events can get lost, so we can't rely on it
to update the "acid" form input data key.

Cancelling the focus events has to nice side effect to not change
the input value when browsing the autocomplete menu.

The value is now update only upon "change" and "select" events,
were we are garantied to have "acid" also updated.