chk_low and chk_high decorators enhancements
Use chk_err value to decide if the cell needs highlighting,
instead of comparing the value with the threshold. The later
method lead to "NaN" being displayed as the treshold value
instead of the proper "empty" value.

It is now perfectly valid to set a low/high threshold value to
None (NULL in db), meaning no threshold.