Unsupport docker repository path with symbolic names for apps and groups
Docker is restrictive about repository names. The validation
regexp is [a-z0-9]+(?:[._-][a-z0-9]+)*

But app codes and group names are not subject to these restrictions,
so they can't be used in repo path elements.

They can not be translated to a docker compatible name either,
because that would risk privilege leaking. For example if we
translate the "foo foo" app name to "foo_foo" and "foo@foo" to
"foo_foo", to /apps/foo_foo would be accessible for the responsibles
of any of the 2 apps.

So just force the app_id and group_id in docker repository paths:

* /apps/<app_id>/<image>
* /groups/<group_id>/<image>

For now keep support /users/<username>/<image>, as the username
is subject to LDAP/AD constraints, and should satisfy the docker
constraints.