Sort sorting outputs
Let the users decide if they want to interleave outputs of
different "Dest".

This also allow the workflow creation at the start of the submit.

The workflow output is also modified to log a "results" data
containing the workflow id and head form id. This data can be
used by later outputs to send a mail linking to the worflow for
example.