Support one more metric data layout
Multi-instance data layout supported the one instance per data
line, using one index to find the instance name and another
index to find the value.

This patch also supports the single line layout, where the instance
is the column name.

This layout parser is activated when either instance or value
index is < 0.