Optimize the server-side form input candidates validation
This validation used to do a full-scan of the candidates to verify the
value is present.

This patch uses the new "search" parameter to fetch only the matching
candidate, if any.

This change aleviate the need for custom api handlers developpers to
implement the same limit=<0|undef> => all results tricks.