[feed] Ping can ask for new daemon status when it detects lost services
When service is deleted in DB but still present on daemon, it will not be recreated until the next
post daemon status that can be never happen in some situations.

We want to detect such situations to force a new post of daemon status.
The most frequent event is daemon ping, so we try to detect during ping worker,
and inform daemon during post daemon ping handler.

## worker ping
    Ping worker walk on svcnames found in latest daemon status data.
    Then find or create services in DB.
    When Created in DB, the service avail status is "undef".
    We use this information set the daemon status required for node:
        HSET <R_DAEMON_STATUS_REQUIRED> <node_id> <timestamp)>

## api ping handler
    if HGET <R_DAEMON_STATUS_REQUIRED> <node_id> exists
        HDEL <R_DAEMON_STATUS_REQUIRED> <node_id>
        returns the resync response: {"ret": 1, "info": "resync", "daemon status required": True})

## daemon collector thread
    When collector daemon ping response is resync, it post new daemon status